Fixed return type of length method to correspond with the Arduino Core

+++ lib/ArduinoNative/src/WString.h
@@ -369,7 +369,7 @@ public:
      *
      * @return String length
      */
-    unsigned int length() const
+    size_t length() const
     {
         unsigned int length = 0;
         const char*  ptr    = m_buffer;
